f32 to f64 and derive debug for Vector3 type

+++ src/mp/mod.rs
@@ -17,28 +17,29 @@ extern "C" {
     pub fn joaat(str: String) -> f64;
 }
 
-#[wasm_bindgen(js_namespace = mp)]
+#[wasm_bindgen(js_namespace = mp, js_name = Vector3)]
 extern "C" {
+    #[derive(Debug)]
     pub type Vector3;
 
     #[wasm_bindgen(constructor)]
-    pub fn new(x: f32, y: f32, z: f32) -> Vector3;
+    pub fn new(x: f64, y: f64, z: f64) -> Vector3;
 
     #[wasm_bindgen(method, getter)]
-    pub fn x(this: &Vector3) -> f32;
+    pub fn x(this: &Vector3) -> f64;
 
     #[wasm_bindgen(method, setter)]
-    pub fn set_x(this: &Vector3, x: f32);
+    pub fn set_x(this: &Vector3, x: f64);
 
     #[wasm_bindgen(method, getter)]
-    pub fn y(this: &Vector3) -> f32;
+    pub fn y(this: &Vector3) -> f64;
 
     #[wasm_bindgen(method, setter)]
-    pub fn set_y(this: &Vector3, y: f32);
+    pub fn set_y(this: &Vector3, y: f64);
 
     #[wasm_bindgen(method, getter)]
-    pub fn z(this: &Vector3) -> f32;
+    pub fn z(this: &Vector3) -> f64;
 
     #[wasm_bindgen(method, setter)]
-    pub fn set_z(this: &Vector3, z: f32);
+    pub fn set_z(this: &Vector3, z: f64);
 }
